shmobile: r8a7740: Set up MERAM address range
The default address range settings have both the MERAM and the IPMMU-PMB
set to use the same range, starting from 0x80000000.  The MERAM has
priority over this address range, which means that regardless of the PMB
setting, any H/W accesses to the 0x80000000 memory region will be intercepted
by the MERAM.  This patch uses the MERAM driver to change the MERAM memory
range to start at 0xC0000000.

Signed-off-by: Damian Hobson-Garcia <dhobsong@igel.co.jp>