ARM: shmobile: r8a7779: Correct TMU clock support
During code review it was noticed that the clock value for
TMU01 was incorrect and the value for TMU02 was missing.

For reference: As of 3.8-rc6 there are no in-tree consumes of these clocks.

Reported-by: Denis Oliver Kropp <dok@directfb.org>
Reviewed-by: Magnus Damm <damm@opensource.se>
Signed-off-by: Simon Horman <horms+renesas@verge.net.au>
(cherry picked from commit 58079fa7d54a0929d304054ee759187a2ccd3cdf)

Conflicts:
	arch/arm/mach-shmobile/clock-r8a7779.c

N.B: commit is in the renesas tree on kernel.org and intended
for inclusion in 3.9.

Signed-off-by: Simon Horman <horms+renesas@verge.net.au>